pyproject.toml
[project]
name = "pdf-to-bw"
version = "0.1.0"
description = "Convert a PDF to black and white"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "typer>=0.24.1",
]


[tool.uv]
package=true

[project.scripts]
retheme-pdf = "retheme_pdf:cli"
